Extract loadState function
