+++ hydra-node/exe/hydra-node/Main.hs
@@ -7,19 +7,13 @@ import Hydra.Prelude
 import Hydra.API.Server (Server (Server, sendOutput), withAPIServer)
 import Hydra.API.ServerOutput (ServerOutput (PeerConnected, PeerDisconnected))
 import Hydra.Cardano.Api (serialiseToRawBytesHex)
-import Hydra.Chain (HeadParameters (..))
 import Hydra.Chain.CardanoClient (QueryPoint (..), queryGenesisParameters)
 import Hydra.Chain.Direct (initialChainState, loadChainContext, mkTinyWallet, withDirectChain)
 import Hydra.Chain.Direct.ScriptRegistry (publishHydraScripts)
 import Hydra.Chain.Direct.Util (readKeyPair)
 import Hydra.HeadLogic (
-  ClosedState (..),
   Environment (..),
   Event (..),
-  HeadState (..),
-  IdleState (..),
-  InitialState (..),
-  OpenState (..),
   defaultTTL,
   getChainState,
  )
@@ -42,6 +36,7 @@ import Hydra.Node (
   HydraNode (..),
   createNodeState,
   initEnvironment,
+  loadState,
   runHydraNode,
  )
 import Hydra.Node.EventQueue (EventQueue (..), createEventQueue)
@@ -49,18 +44,13 @@ import Hydra.Options (
   ChainConfig (..),
   Command (Publish, Run),
   LedgerConfig (..),
-  ParamMismatch (..),
   PublishOptions (..),
   RunOptions (..),
   explain,
   parseHydraCommand,
   validateRunOptions,
  )
-import Hydra.Persistence (Persistence (load), createPersistence, createPersistenceIncremental)
-
-newtype ParamMismatchError = ParamMismatchError String deriving (Eq, Show)
-
-instance Exception ParamMismatchError
+import Hydra.Persistence (createPersistence, createPersistenceIncremental)
 
 main :: IO ()
 main = do
@@ -82,23 +72,7 @@ main = do
         let RunOptions{hydraScriptsTxId, chainConfig} = opts
         -- Load state from persistence or create new one
         persistence <- createPersistence $ persistenceDir <> "/state"
-        hs <-
-          load persistence >>= \case
-            Nothing -> do
-              traceWith tracer CreatedState
-              pure $ Idle IdleState{chainState = initialChainState}
-            Just headState -> do
-              traceWith tracer LoadedState
-              let paramsMismatch = checkParamsAgainstExistingState headState env
-              unless (null paramsMismatch) $ do
-                traceWith tracer (Misconfiguration paramsMismatch)
-                throwIO $
-                  ParamMismatchError $
-                    "Loaded state does not match given command line options."
-                      <> " Please check the state in: "
-                      <> persistenceDir
-                      <> " against provided command line options."
-              pure headState
+        hs <- loadState (contramap Node tracer) env persistence persistenceDir initialChainState
         nodeState <- createNodeState hs
         ctx <- loadChainContext chainConfig party otherParties hydraScriptsTxId
         wallet <- mkTinyWallet (contramap DirectChain tracer) chainConfig
@@ -132,27 +106,6 @@ main = do
     globals <- newGlobals =<< queryGenesisParameters networkId nodeSocket QueryTip
     let ledgerEnv = newLedgerEnv protocolParams
     action (Ledger.cardanoLedger globals ledgerEnv)
-
-  -- check if hydra-node parameters are matching with the hydra-node state.
-  checkParamsAgainstExistingState :: HeadState Ledger.Tx -> Environment -> [ParamMismatch]
-  checkParamsAgainstExistingState hs env =
-    case hs of
-      Idle _ -> []
-      Initial InitialState{parameters} -> validateParameters parameters
-      Open OpenState{parameters} -> validateParameters parameters
-      Closed ClosedState{parameters} -> validateParameters parameters
-   where
-    validateParameters HeadParameters{contestationPeriod = loadedCp, parties} =
-      flip execState [] $ do
-        when (loadedCp /= configuredCp) $
-          modify (<> [ContestationPeriodMismatch{loadedCp, configuredCp}])
-        when (loadedParties /= configuredParties) $
-          modify (<> [PartiesMismatch{loadedParties, configuredParties}])
-     where
-      loadedParties = sort parties
-
-    Environment{contestationPeriod = configuredCp, otherParties, party} = env
-    configuredParties = sort (party : otherParties)
 
 identifyNode :: RunOptions -> RunOptions
 identifyNode opt@RunOptions{verbosity = Verbose "HydraNode", nodeId} = opt{verbosity = Verbose $ "HydraNode-" <> show nodeId}
